The SimH -a and -m flags should help
I've enabled the -m (modem) and -a (autodisconnect) flags in the SimH
configuration. I'm hoping that this won't open up the TCP connection
until it is actually required, and then close it down as needed.

I've done some testing at home and it's OK. I haven't given it a real
workout on Internet sites yet.