Recompiled /vmunix to detect DCD drop, kill off shells after Telnet disconnect.
The GENERIC kernel has flags 0xff on the dz lines. This causes the dz
driver to ignore the DCD drop when a Telnet session disconnects. This
prevents the HUP signal being sent to any shell on that login session,
and so the next telnet session will get a logged-in shell and not a
login prompt.

I've recompiled the kernel with the flags removed.