Fixed the 1000000 problem with pathalias
Seems that pathalias needs to have a link given in both directions in
uucp.map, else it treats it as a dead link. There isn't a flag to
disable this behaviour, so I modified Misc/map2img to notify me of
missing links. Then I edited uucp.map to have the missing links. Now
the pathalias output is good.